requirements.txt
ipython==8.4.0
Faker==13.15.0

#django
Django==4.0.6
django-extensions==3.2.0
django-admin-rangefilter==0.8.5
django-import-export==2.8.0
django-silk==5.0.1
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt


#celery
celery==5.2.7

#parser
beautifulsoup4==4.11.1


#dev
django-debug-toolbar==3.5.0


#flake8
flake8==4.0.1
flake8-debugger==4.1.2
flake8-print==5.0.0
flake8-builtins==1.5.3

#tests
pytest
pytest-django
pytest-cov
pytest-mock

#docs
drf-yasg
